bot/pyproject.toml
[tool.poetry]
name = "test-IT-Lab-bot"
version = "0.1.0"
description = "test-IT-Lab-bot"
authors = ["CodeFramer <52332286+CodeFramer0@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.0"
python-dotenv = "^1.0.1"
environs = "^9.5.0"
marshmallow = "^3.20.1"
pydantic = "^2.11.7"
prometheus-client = "^0.22.1"
aiogram = {version = "2.25.1", extras = ["redis"]}
aioredis = "^2.0.1"
redis = "^6.4.0"
python-dateutil = "^2.9.0.post0"
hashids = "^1.3.1"

[build-system]
requires = ["poetry-core>=1.8.1"]
build-backend = "poetry.core.masonry.api"
